+++ Python/GUI application for photo editing.py
