+++ Python/Take screenshot fun program in python
